Bump up warnings in dev mode.  Fix myriad of warnings from compiling in dev mode
Fix bug with counter not incremented.

Fix potential buffer overrun with vint decoding

git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@10112 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/turbine/worker.c
@@ -38,7 +38,7 @@ static void task_error(Tcl_Interp* interp, int tcl_rc, char* command);
   log_printf("TURBINE_WORKER_SERVICE: " format, ## args)
 
 turbine_code
-turbine_worker_loop(Tcl_Interp* interp, void* buffer, int buffer_size,
+turbine_worker_loop(Tcl_Interp* interp, void* buffer, size_t buffer_size,
                     int work_type)
 {
   int rc;
